Use Jinja2 to template SPARQL query file
We run the SPARQL-anything as a subprocess, passing the SPARQL query
as a file parameter.

However, the SPARQL service paths in the SPARQL query are dynamic.
We need to instantiate the SPARQL file with the correct service paths
and write it in the bag folder.

Use Jinja2 templating to facilitate this.