DEV-2344 Use separate SPARQL files
The SPAQRL file for the basic profile is split up in three separate files.

Rename them to be more descriptive and consistent:
  - descriptive_ie.sparql
  - premis_ie.sparql
  - premis_representation.sparql

Change the transformation logic to use these three SPARQL files. Remove
basic.sparql file.

Corrected a merge issue with the fix for the premis relationship type.

Adapt the empty-graph SIP resources for testing so that the transformed graph
doesn't contain an IE. The SPARQL files are split up meaning that non-matched
IE IDs will result in an IE object in the graph. In short, remove all
references to an IE.

Fix link between representation and the IE for the basic SIP test resources
and remake bags.