DEV-2344 Add fixes to artwork test resources
There should be an UUID identifier used as the identifier for the IE
according to the spec. For the material artwork samples and examples,
add one and fix the links between the metadata files so that an internal
PID is not used as a (linking) identifier.

The local_id with value "VIAA PID" should be "meemoo_pid".

Add extension to the original name.

Remake the bags.