DEV-2344 Add XSD validation for artwork
Validate using the XSD file and append the errors to the list.

Rename `dc_material_artwork.xsd` to `descriptive_material_artwork.xsd`
because the descriptive metadata is extended to also contain schema elements
and attributes. Remove the existing `descriptive_material_artwork.xsd` as this
file was not correct/valid.

Remove print statement.